Move clinics.js to the right place
