+++ app/javascript/src/clinics.js
